Cambios DTO
Se modificaron los cambios que se comentaron en la reunion del jueves 25 de noviembre
