+++ 1dataLayer/1dataLayer.csproj
@@ -53,7 +53,7 @@
     <Compile Include="alergias.cs">
       <DependentUpon>BDCAM.tt</DependentUpon>
     </Compile>
-    <Compile Include="AlexisDTO.cs" />
+    <Compile Include="BDDTO.cs" />
     <Compile Include="alumno.cs">
       <DependentUpon>BDCAM.tt</DependentUpon>
     </Compile>
@@ -78,7 +78,6 @@
     <Compile Include="enfermedades.cs">
       <DependentUpon>BDCAM.tt</DependentUpon>
     </Compile>
-    <Compile Include="fernandoDTO.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="sysdiagrams.cs">
       <DependentUpon>BDCAM.tt</DependentUpon>
